Implementing fixes.
