+++ Dockerfile
@@ -11,8 +11,7 @@ WORKDIR /app
 COPY . .
 RUN go mod download
 
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -gcflags "all=-N -l" cmd/main.go
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o chat-cli -gcflags "all=-N -l" cli/main.go
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o chat -gcflags "all=-N -l" main.go
 
 ##
 ## Deploy
