--- cli/main.go
@@ -1,10 +1,0 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-*/
-package main
-
-import "github.com/daniarmas/chat/cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
