+++ cmd/database/clear.go
