+++ cmd/root.go
@@ -6,8 +6,9 @@ package cmd
 import (
 	"os"
 
-	"github.com/daniarmas/chat/cli/cmd/create"
-	"github.com/daniarmas/chat/cli/cmd/database"
+	"github.com/daniarmas/chat/cmd/create"
+	"github.com/daniarmas/chat/cmd/database"
+	"github.com/daniarmas/chat/cmd/server"
 	"github.com/spf13/cobra"
 )
 
@@ -38,9 +39,23 @@ func Execute() {
 func addSubcommandsPalettes() {
 	rootCmd.AddCommand(database.DatabaseCmd)
 	rootCmd.AddCommand(create.CreateCmd)
+	rootCmd.AddCommand(server.ServerCmd)
 }
 
 func init() {
+	// Here you will define your flags and configuration settings.
+	// Cobra supports persistent flags, which, if defined here,
+	// will be global for your application.
+
+	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.chat.yaml)")
+
+	// Cobra also supports local flags, which will only run
+	// when this action is called directly.
+	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	addSubcommandsPalettes()
+}
+
+func main() {
 	// Here you will define your flags and configuration settings.
 	// Cobra supports persistent flags, which, if defined here,
 	// will be global for your application.
