+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/daniarmas/chat/cmd"
+
+func main() {
+	cmd.Execute()
+}
